pyproject.toml
[project]
name = "home-assistant"
version = "0.1.0"
description = "Voice assistant with wake word detection, speech-to-text, and LLM-powered responses"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "openwakeword>=0.6.0",
    "numpy>=1.24.0",
    "sounddevice>=0.4.6",
    "webrtcvad>=2.0.10",
    "setuptools",
    "python-dotenv",
]

[project.scripts]
assistant = "main:main"

[tool.uv]
dev-dependencies = []
